move easiest of arith methods
