+++ pandas/core/arrays/timedelta.py
@@ -1,8 +1,10 @@
 # -*- coding: utf-8 -*-
 
-from pandas._libs.tslib import Timedelta
+from pandas._libs.tslib import Timedelta, NaT
 
 from pandas.core.dtypes.common import _TD_DTYPE
+
+from pandas.tseries.offsets import Tick
 
 from .datetimelike import DatetimeLikeArrayMixin
 
@@ -15,3 +17,17 @@ class TimedeltaArrayMixin(DatetimeLikeArrayMixin):
     @property
     def dtype(self):
         return _TD_DTYPE
+
+    # ----------------------------------------------------------------
+    # Arithmetic Methods
+
+    def _add_offset(self, other):
+        assert not isinstance(other, Tick)
+        raise TypeError("cannot add the type {typ} to a {cls}"
+                        .format(typ=type(other).__name__,
+                                cls=type(self).__name__))
+
+    def _sub_datelike(self, other):
+        assert other is not NaT
+        raise TypeError("cannot subtract a datelike from a {cls}"
+                        .format(cls=type(self).__name__))
