+++ pandas/core/indexes/datetimelike.py
@@ -696,17 +696,6 @@ class DatetimeIndexOpsMixin(DatetimeLikeArrayMixin):
         # and datetime dtypes
         return self._nat_new(box=True)
 
-    def _sub_nat(self):
-        """Subtract pd.NaT from self"""
-        # GH#19124 Timedelta - datetime is not in general well-defined.
-        # We make an exception for pd.NaT, which in this case quacks
-        # like a timedelta.
-        # For datetime64 dtypes by convention we treat NaT as a datetime, so
-        # this subtraction returns a timedelta64 dtype.
-        # For period dtype, timedelta64 is a close-enough return dtype.
-        result = self._nat_new(box=False)
-        return result.view('timedelta64[ns]')
-
     def _sub_period_array(self, other):
         """
         Subtract one PeriodIndex from another.  This is only valid if they
